+++ packages/antd-management-fast-component/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 2.6.108
+
+### Patch Changes
+
+- update
+
+## 2.6.107
+
+### Patch Changes
+
+- update
+
 ## 2.6.106
 
 ### Patch Changes